Make Amount a Double
Former-commit-id: 56a1fc9e325af73c414e9834f971da9837700f33

+++ src/Data/GS1/EPC.hs
@@ -139,8 +139,7 @@ instance ToSchema SGTINFilterValue
 
 newtype Uom       = Uom {unUom :: T.Text}
   deriving (Show, Read, Eq, Generic, FromJSON, ToJSON)
- -- TODO: Float is almost never the right type
-newtype Amount    = Amount {unAmount :: Float}
+newtype Amount    = Amount {unAmount :: Double}
   deriving (Show, Read, Eq, Generic, FromJSON, ToJSON)
 newtype AssetType = AssetType {unAssetType :: T.Text}
   deriving (Show, Read, Eq, Generic, FromJSON, ToJSON)
